\name{Employment.2}
\docType{data}
\alias{Employment.2}
\title{Employment status in two categories (march dataset format)}
\description{
This dataset contains 845 sequences of 13 observations of a categorical variable representing the professional status categorized into 2 categories: 1="Full time employee", 2="Other situation". 
The first observation of each sequence corresponds to the situation of the respondent at age 20, and then following data were observed each two years, the last observation corresponding to the situation at
age 44. In addition, two covariates are also provided in the dataset. The first one is a fixed covariate representing gender (1="Female", 2="Male"), and the second one is a
time varying covariate representing the health status (1="Good", 2="Bad").
}
\usage{data(Employment.2)}
\format{A march dataset.}
\source{Swiss Household Panel}
\references{
Tillmann, R., Voorpostel, M., Antal, E., Kuhn, U., Lebert, F., Ryser, V.A., Lipps, O., and Wernli, B. (2016). The Swiss Household Panel Study: Observing social change since 1999. Longitudinal and Life Course Studies, 7(1):64-78.

}
\keyword{datasets}