% Generated by roxygen2 (4.0.2): do not edit by hand
\name{march.AIC}
\alias{march.AIC}
\title{Compute Akaike Information Criterion (AIC).}
\usage{
march.AIC(model)
}
\arguments{
\item{model}{The model for which the AIC has to be computed.}
}
\value{
The AIC of the given model.
}
\description{
The AIC (Akaike Information Criterion) is computed for
 a given \code{\link{march.Model-class}} according to the data
 used during construction.
}
\examples{
indepModel <- march.indep.construct(pewee)
march.AIC(indepModel)
}
\author{
Ogier Maitre
}

