% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.diagnostic.R
\name{march.BIC}
\alias{march.BIC}
\title{Compute Bayesian Information Criterion (BIC).}
\usage{
march.BIC(model)
}
\arguments{
\item{model}{The model for which the BIC has to be computed.}
}
\value{
The number of parameters of the given model and its BIC.
}
\description{
The BIC (Bayesian Information Criterion) is computed for
 a given \code{\link{march.Model-class}} according to the data
 used during construction.
}
\examples{
indepModel <- march.indep.construct(pewee)
march.BIC(indepModel)
}
\author{
Ogier Maitre
}
