% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.AllClasses.R
\docType{class}
\name{march.Dataset-class}
\alias{march.Dataset-class}
\title{Dataset for march package.}
\description{
This class contains several discrete-valued time series, in a dataset. It contains for each sequence, 
its length and weights.
}
\details{
The internal representation uses factor-like representation. The integer values correspond to the words stored
into the dictionary vector. Therefor, they are in the interval [1,K].

 @section Slots:
 \describe{
   \item{\code{yRaw}:}{A matrix of \code{\link{character}} string, describing the content of 
   the original dataset or file, if any.}
   \item{\code{y}:}{A list of vector of \code{\link{integer}} representing the  each discrete-valued time series
   of the dataset, as can be used by the models.}
   \item{\code{T}:}{A vector of \code{\link{integer}} values representing the length of each sequence.}
   \item{\code{weights}:}{A vector of \code{\link{numeric}} values representing the weight of each sequence.}
   \item{\code{K}:}{A \code{\link{integer}} value representing the number of possible ouput and the number
   of words stored into the dictionary.}
   \item{\code{N}:}{A \code{\link{integer}} value representing the number of sequence.}
   \item{\code{Dictionary}:}{A vector of \code{\link{character}} string representing the translation between 
   the yRaw and y data. Each character string is stored according to the integer which represents it into y.}
   \item{\code{cov}:}{A matrix of \code{\link{integer}} representing the covariates.}
   \item{\code{Kcov}:}{A vector of \code{\link{integer}} representing the number of possible output for each covariate.}
   \item{\code{Ncov}:}{A \code{\link{integer}} value representing the number of covariates.}
 }
 @seealso \code{\link{march.dataset.loadFromFile}}, \code{\link{march.dataset.loadFromDataFrame}}
 @author Ogier Maitre
}
