% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.dataset.R
\name{march.dataset.h.extractSequence}
\alias{march.dataset.h.extractSequence}
\title{Extract a sequence from a dataset.}
\usage{
march.dataset.h.extractSequence(y, i)
}
\arguments{
\item{y}{is the dataset from which one sequence of data is extracted}

\item{i}{is the number of sequence to extract}
}
\description{
Extract a sequence from a dataset.
}

