% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.AllClasses.R
\docType{class}
\name{march.Indep-class}
\alias{march.Indep-class}
\title{An independence model.}
\description{
This class describes an independence model, represented by the probability distribution \emph{indP} of each event
and the number of data used to compute each member of the probability distribution. march.Indep inherits
from \code{\link{march.Model-class}} and therefore inherits its slots.
}
\section{Slots}{

 \describe{
   \item{\code{indP}:}{A vector of \code{\link{numeric}} representing the model probability distribution.}
   \item{\code{indC}:}{A vector of \code{\link{integer}} representing the number of data used to compute each
   member of the probability distribution.}
   }
}

\seealso{
\code{\link{march.indep.construct}}, \code{\link{march.Model-class}}.
}
