% Generated by roxygen2 (4.0.2): do not edit by hand
\name{march.indep.construct}
\alias{march.indep.construct}
\title{Construct an independence model.}
\usage{
march.indep.construct(y, maxOrder = 0)
}
\arguments{
\item{y}{the \code{\link{march.Dataset}} from which construct the model.}

\item{maxOrder}{the maximum visible order among the set of Markovian models to compare.}
}
\value{
The \code{\link{march.Indep}} constructed using dataset y and maxOrder.
}
\description{
Construct a \code{\link{march.Indep}} model from a given \code{\link{march.Dataset}},
the first \emph{maxOrder} elements of each sequence being truncated in order to return a model
which can be compared with other Markovian models of visible order maxOrder.
}
\examples{
# Build an independance model from the pewee data set.
model <- march.indep.construct(pewee)
print(model)
}
\author{
Ogier Maitre
}
\seealso{
\code{\link{march.Indep}}, \code{\link{march.Model}}, \code{\link{march.Dataset}}.
}

