% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.AllClasses.R
\docType{class}
\name{march.Mc-class}
\alias{march.Mc-class}
\title{A Markov chain of order >= 1.}
\description{
This class describes a Markov chain of order \emph{order}, represented by matricess RC (transition matrix in reduced form)
and RT (number of data points used to compute each transition). march.Mc extends \code{\link{march.Model-class}} class and therefore
inherits its slots.
}
\section{Slots}{

 \describe{
   \item{\code{RC}:}{A matrix of \code{\link{numeric}} representing the reduced form of the 
   transition matrix of the current Markov Chain.}
   \item{\code{order}:}{An \code{\link{integer}} representing the order of the current Markov Chain.}
   \item{\code{RT}:}{A matrix of \code{\link{integer}} representing the number of sample used to compute each 
   transition row of the current RC matrix.}
 }
}

\seealso{
\code{\link{march.mc.construct}}, \code{\link{march.Model-class}}.
}
