% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.AllClasses.R
\docType{class}
\name{march.Model-class}
\alias{march.Model-class}
\title{A basic and virtual march model.}
\description{
This class describe the basic and virtual model, that every model of the package will extend.
This is a virtual class, which is not meant to be handled by user directly.
}
\seealso{
The classes that inherit from march.Model are : \code{\link{march.Indep-class}}, \code{\link{march.Mc-class}}, \code{\link{march.Mtd-class}}, \code{\link{march.Dcmm-class}}.

 @section Slots:
   \describe{
     \item{\code{ll}:}{A \code{\link{numeric}} representing the log-likelihood for this model \emph{w.r.t} its 
     construction dataset.}
     \item{\code{y}:}{The \code{\link{march.Dataset-class}} used to construct the model.}
     \item{\code{dsL}:}{A \code{\link{numeric}} representing the number of sample used to construct the model.}
     \item{\code{nbZeros}:}{A \code{\link{numeric}} representing the number of zeros created during model construction.}
     }
}
