% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.AllGenerics.R
\docType{methods}
\name{march.name,march.Mtd-method}
\alias{march.name,march.Mtd-method}
\title{This method is called with the object "march.Mtd" and provides it 
to the march.name function.}
\usage{
\S4method{march.name}{march.Mtd}(object)
}
\arguments{
\item{object}{contains the name of the model.}
}
\description{
This method is called with the object "march.Mtd" and provides it 
to the march.name function.
}

