% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.AllGenerics.R
\name{march.name}
\alias{march.name}
\title{march.Model name.}
\usage{
march.name(object)
}
\description{
Generate a name for the march model contained in the given \emph{object}.
}
\author{
Ogier Maitre & Andre Berchtold
}
