\name{march-package}
\alias{march}
\title{Computation of Markovian models for categorical data}
\description{
This package is dedicated to the computation of various Markovian models for categorical data including the independence model, homogeneous Markov chains of any order, the Mixture Transition Distribution (MTD) model for the approximation of high-order homogeneous Markov chains, Hidden Markov Models (HMMs) and Double Chain Markov Models (DCMMs).
}
\author{Ogier Maitre and Kevin Emery, with contributions from Oliver Buschor and Andre Berchtold}
\references{
 Berchtold A, Raftery AE (2002) The Mixture Transition Distribution Model for High-Order Markov Chains and Non-Gaussian Time-Series.
 \emph{Statistical Science} 17(3), 328-356.
 
 Berchtold A (2002) High-order extensions of the Double Chain Markov Model. 
 \emph{Stochastic Models} 18, 193-227.
}

\seealso{
\code{\link{march.Model-class}}, \code{\link{march.Dataset-class}}.
}