% Generated by roxygen2 (4.0.2): do not edit by hand
\name{march.read}
\alias{march.read}
\title{Load a march.Model.}
\usage{
march.read(filename)
}
\arguments{
\item{filename}{the path where load the mode}
}
\value{
the march.Model contained into the file pointed by filename if it exists and contains a valid model.
}
\description{
Load a march.Model from a file pointed by \emph{filename} and check that the model is valid.
}

