% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.AllGenerics.R
\name{march.summary}
\alias{march.summary}
\title{march.Model Summary.}
\usage{
march.summary(object, ...)
}
\arguments{
\item{object}{can contain the results of any model computed  using march}

\item{...}{should indicate any additional parameter passed to the function}
}
\description{
Print key values for the current model.
}
\author{
Ogier Maitre & Andre Berchtold
}
