% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.AllGenerics.R
\docType{methods}
\name{march.thompson,march.Mtd,numeric-method}
\alias{march.thompson,march.Mtd,numeric-method}
\title{This method is called with the object "march.Mtd" and the aplha "numeric" and
provides it to the march.thompson function.}
\usage{
\S4method{march.thompson}{march.Mtd,numeric}(object, alpha)
}
\arguments{
\item{object}{contains the name of the model.}

\item{alpha}{contains the Type I error}
}
\description{
This method is called with the object "march.Mtd" and the aplha "numeric" and
provides it to the march.thompson function.
}

