\name{pewee_df}
\docType{data}
\alias{pewee_df}
\title{Song of the Wood Pewee (data frame format)}
\description{
This dataset contains a sequence of 1327 successive observations of the wood pewee song. This song consists in three different phrases numbered from 1 to 3.
}
\usage{pewee_df}
\format{A data frame.}
\source{Craig (1943)}
\references{
Craig, W. (1943) \emph{The Song of the Wood Peewee}; University of the State of New York:
Albany.
}
\keyword{datasets}