\name{sleep}
\docType{data}
\alias{sleep}
\title{Sleep disorders (march dataset format)}
\description{
This dataset contains 1000 sequences of 5 to 7 successive observation of the level of sleep disorders. Each sequence corresponds to a different subject. Possible values range from 1 (no disorder at all) to 6 (disturbed sleep each night). 
}
\usage{data(sleep)}
\format{A march dataset.}
\keyword{datasets}