\name{sleep_df}
\docType{data}
\alias{sleep_df}
\title{Sleep disorders (data frame format)}
\description{
This dataset contains 1000 sequences of 5 to 7 successive observation of the level of sleep disorders. Each sequence corresponds to a different subject. Possible values range from 1 (no disorder at all) to 6 (disturbed sleep each night). 
}
\usage{data(sleep_df)}
\format{A data frame.}
\keyword{datasets}