% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.diagnostic.R
\name{march.AIC}
\alias{march.AIC}
\title{Compute Akaike Information Criterion (AIC).
 
 The AIC (Akaike Information Criterion) is computed for
 a given \code{\link{march.Model-class}} according to the data
 used during construction.}
\usage{
march.AIC(model)
}
\arguments{
\item{model}{The model for which the AIC has to be computed.}
}
\value{
The number of parameters of the given model and its AIC.
}
\description{
Compute Akaike Information Criterion (AIC).
 
 The AIC (Akaike Information Criterion) is computed for
 a given \code{\link{march.Model-class}} according to the data
 used during construction.
}
\examples{
indepModel <- march.indep.construct(pewee)
march.AIC(indepModel)
}
\author{
Ogier Maitre
}
