% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.dataset.R
\name{march.dataset.h.extractSequence}
\alias{march.dataset.h.extractSequence}
\title{Extract a sequence from a dataset.}
\usage{
march.dataset.h.extractSequence(y, i)
}
\arguments{
\item{y}{A sequence of integers.}

\item{i}{The number of observations to keep.}
}
\description{
Extract a sequence from a dataset.
}
\author{
Ogier Maitre
}
