% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.dataset.R
\name{march.dataset.loadFromDataFrame}
\alias{march.dataset.loadFromDataFrame}
\title{Construct a dataset from a data.frame or a matrix.}
\usage{
march.dataset.loadFromDataFrame(dataframe, MARGIN = 2, weights = NA,
  missingDataRep = NA, covariates = NULL)
}
\arguments{
\item{dataframe}{A \code{\link{data.frame}} containing the dataset.}

\item{MARGIN}{The dimension of the matrix/data.frame that contains the sequences (resp 1 for the column, 2 for the rows).}

\item{weights}{If specified, contains the weight of each sequence.}

\item{missingDataRep}{If specified, the symbol representing a missing data.}
}
\value{
A \code{\link{march.Dataset-class}} object containing the data contructed from the matrix or data.frame.
}
\description{
The function creates a \code{\link{march.Dataset}} from a \emph{dataframe} or a \emph{matrix}, where each row (resp. column) represents 
an independent data series when \emph{MARGIN} is 2 (resp. 1).
}
\author{
Ogier Maitre
}
