% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.dcmm.R
\name{march.dcmm.construct}
\alias{march.dcmm.construct}
\title{Construct a double chain Markov model (DCMM).}
\usage{
march.dcmm.construct(y, orderHC, orderVC, M, gen = 5, popSize = 4,
  maxOrder = orderVC, seedModel = NULL, iterBw = 2, stopBw = 0.1)
}
\arguments{
\item{y}{the dataset from which the Dcmm will be constructed \code{\link{march.Dataset}}.}

\item{orderHC}{the order of the hidden chain of the constructed Dcmm.}

\item{orderVC}{the order of the visible chain of the constructed Dcmm (0 for a HMM).}

\item{M}{the number of hidden state of the Dcmm.}

\item{gen}{the number of generation performed by the EA.}

\item{popSize}{the number of individual stored into the population.}

\item{maxOrder}{the maximum visible order among the set of Markovian models to compare.}

\item{seedModel}{a model to optimize using Baum-Welch algorithm.}

\item{iterBw}{the number of iteration performed by the Baum-Welch algorithm.}

\item{stopBw}{the minimum increase in quality (log-likelihood) authorized in the Baum-Welch algorithm.}
}
\value{
the best \code{\link{march.Dcmm}} constructed by the EA or the result of the Baum-Welch algorithm on \emph{seedModel}.
}
\description{
Construct a \code{\link{march.Dcmm}} object, with visible order \emph{orderVC}, hidden order \emph{orderHC} and \emph{M} hidden states, according to a \code{\link{march.Dataset}}.
The first \emph{maxOrder}-\emph{orderVC} elements of each sequence are truncated in order to return a model
which can be compared with other Markovian model of visible order maxOrder. The construction is performed either by an evolutionary algorithm (EA) or by improving an existing DCMM.
The EA performs \emph{gen} generations on a population of \emph{popSize} individuals. The EA behaves as a Lamarckian evolutionary algorithm, using a Baum-Welch algorithm as
optimization step, running until log-likelihood improvement is less than \emph{stopBw} or for \emph{iterBw} iterations. Finally only the best individual from the population is returned as solution.
If a seedModel is provided, the only step executed is the optimization step, parameters related to the EA does not apply in this case.
}
\seealso{
\code{\link{march.Dcmm-class}}, \code{\link{march.Model-class}}, \code{\link{march.Dataset-class}}.
}
\author{
Ogier Maitre
}
