% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.AllGenerics.R
\name{march.name}
\alias{march.name}
\title{march.Model name.}
\usage{
march.name(object)
}
\arguments{
\item{object}{the (\code{\link{march.Model-class}}) from which to extract the name.}
}
\description{
Generate a name for the march model contained in the given \emph{object}.
}
\examples{
# Compute a third-order homogeneous Markov Chain for the pewee data.
model <- march.mc.construct(pewee,3)

# Assign its name to the model.
march.name(model)
}
\author{
Ogier Maitre & Andre Berchtold
}
