% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march.AllGenerics.R
\name{march.summary}
\alias{march.summary}
\title{march.Model Summary.}
\usage{
march.summary(object)
}
\arguments{
\item{object}{the (\code{\link{march.Model-class}}) from which to extract the informations.}
}
\description{
Print the log-likelihood, the number of non-zeros parameter, the AIC and the BIC of the current model.
}
\seealso{
\code{\link{march.Model-class}}
}
\author{
Ogier Maitre & Andre Berchtold
}
